+++ blueprints/operating-entity/multi-operating-entities/providers.tf
@@ -4,3 +4,10 @@ provider "oci" {
   region              = var.region
   config_file_profile = var.oci_config_profile
 }
+
+provider "oci" {
+  alias               = "home"
+  tenancy_ocid        = var.tenancy_ocid
+  region              = coalesce(var.home_region, var.region)
+  config_file_profile = var.oci_config_profile
+}